reset password, resend email

+++ src/auth/model/verification-tokens/verify-token.ts
@@ -1,5 +1,5 @@
-import { PrismaClient } from "@prisma/client";
-import {verify} from "argon2";
+import { $Enums, PrismaClient } from '@prisma/client';
+import { verify } from 'argon2';
 
 /**
  * Verify a user’s token
@@ -10,23 +10,23 @@ import {verify} from "argon2";
  * @param prisma PrismaClient - Injected Prisma client
  */
 export async function VerifyToken(
-  scope: string,
+  scope: $Enums.VerificationTokenScope,
   tokenId: string,
   otp: string,
-  prisma: PrismaClient
+  prisma: PrismaClient,
 ) {
   const token = await prisma.verificationToken.findUnique({
     where: { id: tokenId },
   });
 
-  if (!token) return { error: "Invalid or expired token." };
-  if (token.expires_at < new Date()) return { error: "Token expired." };
-
+  if (!token) return { error: 'Invalid or expired token.' };
+  if (token.expires_at < new Date()) return { error: 'Token expired.' };
+  if (token.scope !== scope) return { error: 'Un authorized token' };
   const isValid = await verify(token.token, otp);
-  if (!isValid) return { error: "Invalid OTP." };
+  if (!isValid) return { error: 'Invalid OTP.' };
 
   // Optional cleanup: delete token after success
   await prisma.verificationToken.delete({ where: { id: tokenId } });
 
-  return { success: true, userid:token.identifier_id};
+  return { success: true, userid: token.identifier_id };
 }
